Move ExpirationCommand to a generic util
